Cleanup bugs; more complete testing

+++ install_R.bash
@@ -18,33 +18,48 @@ trap f ERR
 
 R_VERSION=$1
 R_BASE=${2:-R-3}
-# download R
-mkdir /usr/local/src/R
-cd /usr/local/src/R
-wget --no-check-certificate https://cran.fhcrc.org/src/base/$R-BASE/R-$R_VERSION.tar.gz
-tar zxf R-$R_VERSION.tar.gz
+if [ ! -f /usr/local/R-$R_VERSION/lib/R/bin/R ]; then
+    echo "Building R-$R_VERSION ..."
+    # download R
+    if [ ! -d /usr/local/src/R ]; then
+        mkdir /usr/local/src/R
+    fi
+    cd /usr/local/src/R
 
-# create the R target install
-mkdir /usr/local/R-$R_VERSION
+    if [ ! -f R-$R_VERSION.tar.gz ]; then
+        wget --no-check-certificate https://cran.fhcrc.org/src/base/$R_BASE/R-$R_VERSION.tar.gz
+    fi
+    tar zxf R-$R_VERSION.tar.gz
+    cd /usr/local/src/R/R-$R_VERSION
 
-# set config vars for building R; config; make
-MKL_LIB_PATH=/opt/intel/mkl/lib/intel64
-export LD_LIBRARY_PATH=$MKL_LIB_PATH
-MKL="-m64 -L${MKL_LIB_PATH} -lmkl_gf_lp64 -lmkl_core -lmkl_sequential -lpthread -lm"
-cd /usr/local/src/R/R-$R_VERSION
-./configure --enable-R-shlib --enable-threads=posix --prefix=/usr/local/R-3.3.2 --with-blas="$MKL" --enable-memory-profiling
-make
-make check
-make info
-make install
+    # create the R target install
+    if [ ! -d /usr/local/R-$R_VERSION ]; then
+        mkdir /usr/local/R-$R_VERSION
+    fi
 
-# create R links
-ln -s /usr/local/R-$R_VERSION/lib/R/bin/R /usr/local/bin/R
-ln -s /usr/local/R-$R_VERSION/lib/R/bin/Rscript /usr/local/bin/Rscript
+    # set config vars for building R; config; make
+    MKL_LIB_PATH=/opt/intel/mkl/lib/intel64
+    export LD_LIBRARY_PATH=$MKL_LIB_PATH
+    MKL="-m64 -L${MKL_LIB_PATH} -lmkl_gf_lp64 -lmkl_core -lmkl_sequential -lpthread -lm"
+    ./configure --enable-R-shlib --enable-threads=posix --prefix=/usr/local/R-3.3.2 --with-blas="$MKL" --enable-memory-profiling
+    make
+    make check
+    make info
+    make install
 
-# install RStudio
-sudo apt-get install gdebi-core
-mkdir /usr/local/src/rstudio
-cd /usr/local/src/rstudio
-wget https://download2.rstudio.org/rstudio-server-1.0.143-amd64.deb
-sudo dpkg -i rstudio-server-1.0.143-amd64.deb
+    # create R links
+    if [ ! -d /usr/local/bin/R ]; then
+        ln -s /usr/local/R-$R_VERSION/lib/R/bin/R /usr/local/bin/R
+    else
+        rm /usr/local/bin/R
+        ln -s /usr/local/R-$R_VERSION/lib/R/bin/R /usr/local/bin/R
+    fi
+    if [ ! -d /usr/local/bin/Rscript ]; then
+        ln -s /usr/local/R-$R_VERSION/lib/R/bin/Rscript /usr/local/bin/Rscript
+    else
+        rm /usr/local/bin/Rscript
+        ln -s /usr/local/R-$R_VERSION/lib/R/bin/Rscript /usr/local/bin/Rscript
+    fi
+else
+    echo "R-$R_VERSION already built"
+fi
